refactor: split out reshape subchunk grid stuff and document
